AutoBloggerApp/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
PyQt6==6.6.1
PyQt5>=5.15.0
PyQtWebEngine>=5.15.0
pyqtgraph>=0.13.0
mistralai>=0.0.7
openai>=1.3.7
python-wordpress-xmlrpc>=2.3
tweepy>=4.14.0
linkedin-api==2.0.0a5
python-linkedin-v2>=0.8.3
cryptography>=41.0.5

# Data processing and analysis
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Web and API
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1

# Testing and development
pytest>=7.0.0
pytest-cov==4.1.0
flake8>=6.0.0
black==23.11.0
mypy==1.7.1

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
rich==13.7.0
loguru==0.7.2

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Security
cryptography==41.0.5
python-jose==3.3.0
passlib==1.7.4

# BeautifulSoup4
beautifulsoup4>=4.12.2
markdown>=3.3.0

# Selenium
selenium>=4.10.0
selenium-stealth>=1.0.6
undetected-chromedriver>=3.5.0

# Webdriver Manager
webdriver-manager>=4.0.0 